Fix return-type mismatch in FP6 dequantize helpers
`convert_e2m3x2_to_fp16` and `convert_e3m2x2_to_fp16` have two branches
selected by the `MAJOR_COMPUTE_CAPABILITY` constexpr argument:

  - The Blackwell branch uses inline PTX (`cvt.rn.f16x2.e{2m3,3m2}x2`)
    and returns `tl.float16`.
  - The fallback calls `convert_low_precision_fp_to_high_precision`,
    which ends in `.to(tl.float32)` and therefore returns `tl.float32`.

Even though `MAJOR_COMPUTE_CAPABILITY` is `tl.constexpr`, Triton's
AST-to-TTIR type inference walks both branches and requires every
`return` path of a `@triton.jit` function to share a dtype. Triton
3.4, 3.5, 3.6, and 3.7 all reject these helpers at compile time with

    Inconsistent return types: <[...], fp16> and <[...], fp32>

which makes `dequantize` unusable for every dtype routed through them:
nvfp6_e2m3, nvfp6_e3m2, mxfp6_e2m3, mxfp6_e3m2 (directly, via
`kernels/triton/dequantize.py`), and if6_e2m3, if6_e3m2 (indirectly,
via `kernels/triton/if6.py`).

Cast the fallback result to `tl.float16` so both branches agree. The
quantize-side counterparts `convert_to_e{2m3,3m2}x2` are unaffected
because both of their branches already return `uint8`.

The bug is not caught by CI: `test_backend_outputs_are_consistent`
only compares quantized values and scale factors across backends and
never invokes `dequantize`, while `test_pseudo_quantize` is restricted
to 4-bit dtypes. The FP6 / IF6 dequantize paths have no automated
coverage today.

Minimal repro on a Blackwell GPU:

    import torch
    from fouroversix import (
        QuantizationConfig, quantize, dequantize,
        DataType, QuantizeBackend, ScaleRule,
    )
    x = torch.randn(128, 128, dtype=torch.bfloat16, device="cuda")
    cfg = QuantizationConfig(
        backend=QuantizeBackend.triton,
        dtype=DataType.nvfp6_e2m3,
        scale_rule=ScaleRule.static_6,
    )
    qt = quantize(x, cfg)        # passes
    dequantize(qt)               # CompilationError before this patch

Verified on B200 (sm_100) with Triton 3.6.0: after the patch all six
previously-broken dtypes round-trip cleanly and produce bit-exact
output relative to the PyTorch reference backend on 1024x1024 inputs
(RMSE 0.026 for nvfp6_e2m3 / static_6 and 0.047 for nvfp6_e3m2 /
static_6 against the original bfloat16 tensor).

+++ src/fouroversix/kernels/triton/fp6.py
@@ -87,7 +87,7 @@ def convert_e3m2x2_to_fp16(
         EXP_BIAS=1,
         EXP_BITS=3,
         MANTISSA_BITS=2,
-    )
+    ).to(tl.float16)
 
 
 @triton.jit
@@ -261,4 +261,4 @@ def convert_e2m3x2_to_fp16(
         EXP_BIAS=1,
         EXP_BITS=2,
         MANTISSA_BITS=3,
-    )
+    ).to(tl.float16)
